fix:  floor 이미지 업데이트시 모든 씬 재구성되는 로직 수정
- update 클로저에서 image URL비교해서 처리
