+++ SpacialMoodBoard/Sources/Presentations/SceneRealityView/ViewModel/SceneViewModel.swift
@@ -50,6 +50,9 @@ final class SceneViewModel {
     // Floor 이미지 선택 모드
     var isSelectingFloorImage: Bool = false
     
+    // Floor에 적용된 이미지 URL
+    var appliedFloorImageURL: URL?
+    
     // SceneObjects (computed property)
     var sceneObjects: [SceneObject] {
         guard let scene = appStateManager.selectedScene else { return [] }
@@ -92,45 +95,7 @@ final class SceneViewModel {
         attachmentTimerTask?.cancel()
         attachmentTimerTask = nil
     }
-    
-    // MARK: - Floor Material Management
-    
-    func applyFloorImage(from asset: Asset) {
-        guard asset.type == .image else {
-            return
-        }
-        
-        // Documents 디렉토리로부터의 상대 경로 계산
-        let relativePath: String?
-        if let documentsURL = FileManager.default.urls(for: .documentDirectory, in: .userDomainMask).first {
-            let documentsPathWithSlash = documentsURL.path + "/"
-            if asset.url.path.hasPrefix(documentsPathWithSlash) {
-                relativePath = String(asset.url.path.dropFirst(documentsPathWithSlash.count))
-            } else {
-                print("⚠️ Asset이 Documents 디렉토리 내에 없음: \(asset.url.path)")
-                relativePath = nil
-            }
-        } else {
-            print("⚠️ Documents 디렉토리를 찾을 수 없음")
-            relativePath = nil
-        }
-        
-        // SpacialEnvironment에 floor material URL과 상대 경로 저장
-        var updatedEnvironment = spacialEnvironment
-        updatedEnvironment.floorMaterialImageURL = asset.url
-        updatedEnvironment.floorImageRelativePath = relativePath
-        spacialEnvironment = updatedEnvironment
 
-        // Floor entity 캐시 초기화 (다음 호출 시 새 material로 재생성됨)
-        entityRepository.clearFloorCache()
-
-        // 선택 모드 해제
-        isSelectingFloorImage = false
-        
-        // 변경사항 저장
-        saveScene()
-    }
-    
     // MARK: - Scene Persistence
     
     /// SceneModel을 디스크에 저장
